Init snippy update to v4
