+++ .gitignore
@@ -7,3 +7,6 @@ target
 /.settings
 /bin
 *.factorypath
+*.classpath
+*.project
+*.settings
