--- irida-sanbi-bvc-pipeline-plugin/src/main/resources/workflows/0.1.0/messages_en.properties
@@ -1,61 +1,0 @@
-#Pipeline Info Properties
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.title.SanbiBVCPipeline=Pipelines - Sanbi Bacterial Variant Calling
-workflow.phylogenomics.description=The pipeline employs the Genome Analysis Toolkit (GATK) to perform variant calling and is based on the best practices for variant discovery analysis outlined by the Broad Institute. Once SNPs have been identified, SnpEff is utilized to annotate and predict the effects of the variants
-pipeline.parameters.modal-title.sanbibvcpipeline=Sanbi Bacterial Variant Calling Pipeline Parameters
-pipeline.h1.SanbiBVCPipeline=SanbiBVCPipeline Pipeline
-workflow.phylogenomics.title=SanbiBVCPipeline Pipeline
-#Tool Parameters - Tool: trimmomatic - Workflow Step #: 2
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.trimmomatic-2-illuminaclip.do_illuminaclip=trimmomatic-2-illuminaclip.do_illuminaclip
-pipeline.parameters.sanbibvcpipeline.trimmomatic-2-readtype.single_or_paired=trimmomatic-2-readtype.single_or_paired
-#Tool Parameters - Tool: novo_align - Workflow Step #: 3
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.novo_align-3-paired_end.paired_input_type_conditional.paired_input_type=novo_align-3-paired_end.paired_input_type_conditional.paired_input_type
-pipeline.parameters.sanbibvcpipeline.novo_align-3-paired_end.is_paired_end=novo_align-3-paired_end.is_paired_end
-pipeline.parameters.sanbibvcpipeline.novo_align-3-index.source=novo_align-3-index.source
-#Tool Parameters - Tool: picard - Workflow Step #: 4
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.picard-4-read_name_regex=picard-4-read_name_regex
-pipeline.parameters.sanbibvcpipeline.picard-4-optical_duplicate_pixel_distance=picard-4-optical_duplicate_pixel_distance
-pipeline.parameters.sanbibvcpipeline.picard-4-duplicate_scoring_strategy=picard-4-duplicate_scoring_strategy
-pipeline.parameters.sanbibvcpipeline.picard-4-validation_stringency=picard-4-validation_stringency
-pipeline.parameters.sanbibvcpipeline.picard-4-remove_duplicates=picard-4-remove_duplicates
-pipeline.parameters.sanbibvcpipeline.picard-4-assume_sorted=picard-4-assume_sorted
-#Tool Parameters - Tool: picard - Workflow Step #: 5
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.picard-5-bisulphite=picard-5-bisulphite
-pipeline.parameters.sanbibvcpipeline.picard-5-assume_sorted=picard-5-assume_sorted
-pipeline.parameters.sanbibvcpipeline.picard-5-maxinsert=picard-5-maxinsert
-pipeline.parameters.sanbibvcpipeline.picard-5-reference_source.reference_source_selector=picard-5-reference_source.reference_source_selector
-pipeline.parameters.sanbibvcpipeline.picard-5-validation_stringency=picard-5-validation_stringency
-#Tool Parameters - Tool: freebayes - Workflow Step #: 6
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.freebayes-6-target_limit_type.target_limit_type_selector=Limit variant calling to a set of regions?
-pipeline.parameters.sanbibvcpipeline.freebayes-6-reference_source.reference_source_selector=Load reference genome from
-pipeline.parameters.sanbibvcpipeline.freebayes-6-options_type.options_type_selector=Choose parameter selection level
-#Tool Parameters - Tool: qualimap2 - Workflow Step #: 7
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.qualimap2-7-mem_size=qualimap2-7-mem_size
-#Tool Parameters - Tool: snpeff - Workflow Step #: 8
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.snpeff-8-outputConditional.outputFormat=Output format
-pipeline.parameters.sanbibvcpipeline.snpeff-8-spliceRegion.setSpliceRegions=spliceRegion Settings
-pipeline.parameters.sanbibvcpipeline.snpeff-8-udLength=Upstream / Downstream length
-pipeline.parameters.sanbibvcpipeline.snpeff-8-offset=Chromosomal position
-pipeline.parameters.sanbibvcpipeline.snpeff-8-filter.specificEffects=Filter out specific Effects
-pipeline.parameters.sanbibvcpipeline.snpeff-8-spliceSiteSize=Set size for splice sites (donor and acceptor) in bases
-pipeline.parameters.sanbibvcpipeline.snpeff-8-snpDb.genome_version=Snpff Genome Version Name (e.g. GRCh38.76)
-pipeline.parameters.sanbibvcpipeline.snpeff-8-chr=Text to prepend to chromosome name
-pipeline.parameters.sanbibvcpipeline.snpeff-8-generate_stats=Produce Summary Stats
-pipeline.parameters.sanbibvcpipeline.snpeff-8-noLog=Suppress reporting usage statistics to server
-pipeline.parameters.sanbibvcpipeline.snpeff-8-snpDb.genomeSrc=Genome source
-pipeline.parameters.sanbibvcpipeline.snpeff-8-inputFormat=Input format
-#Tool Parameters - Tool: vcf_to_alignment - Workflow Step #: 9
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.vcf_to_alignment-9-remove_invariant_sites=vcf_to_alignment-9-remove_invariant_sites
-pipeline.parameters.sanbibvcpipeline.vcf_to_alignment-9-reference.source=vcf_to_alignment-9-reference.source
-#Tool Parameters - Tool: fasttree - Workflow Step #: 10
-#Mon Mar 11 12:28:25 SAST 2019
-pipeline.parameters.sanbibvcpipeline.fasttree-10-advanced_options.advanced_options_selector=fasttree-10-advanced_options.advanced_options_selector
-pipeline.parameters.sanbibvcpipeline.fasttree-10-advanced_options.alignment_type1=fasttree-10-advanced_options.alignment_type1
