--- irida-snippy-pipeline-plugin/pom.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>ca.corefacility.bioinformatics.irida.plugins</groupId>
-	<artifactId>irida-snippy-pipeline-plugin</artifactId>
-	<version>0.1.1-SNAPSHOT</version>
-	<!-- Please fill out these properties with information about your particular plugin -->
-	<properties>
-		<!-- Information used to define properties about a plugin. Please see the PF4J docs for more details https://pf4j.org/doc/getting-started.html -->
-		<plugin.id>irida-snippy-pipeline-plugin</plugin.id>
-		<plugin.class>ca.corefacility.bioinformatics.irida.plugins.SnippyBVCPipelinePlugin</plugin.class>
-		<plugin.version>0.1.1</plugin.version>
-		<plugin.provider>SANBI-SA</plugin.provider>
-		<plugin.dependencies></plugin.dependencies>
-		<plugin.requires.runtime>1.0.0</plugin.requires.runtime>
-		<!-- Version of the pf4j dependency -->
-		<pf4j.version>2.4.0</pf4j.version>
-		<!-- Version of Java to target for this plugin -->
-		<java.version>1.8</java.version>
-		<!-- The version of IRIDA your plugin must be compiled against. -->
-		<irida.version.compiletime>19.04-SNAPSHOT</irida.version.compiletime>
-	</properties>
-	<!-- Dependencies used to build this plugin package -->
-	<dependencies>
-		<dependency>
-			<groupId>ca.corefacility.bioinformatics</groupId>
-			<artifactId>irida</artifactId>
-			<version>${irida.version.compiletime}</version>
-			<classifier>classes</classifier>
-		</dependency>
-		<dependency>
-			<groupId>org.pf4j</groupId>
-			<artifactId>pf4j</artifactId>
-			<version>${pf4j.version}</version>
-			<scope>provided</scope>
-		</dependency>
-	</dependencies>
-	<!-- You should normally not have to modify any of the below information -->
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.1</version>
-				<configuration>
-					<source>${java.version}</source>
-					<target>${java.version}</target>
-					<forceJavacCompilerUse>true</forceJavacCompilerUse>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-jar-plugin</artifactId>
-				<version>3.1.1</version>
-				<configuration>
-					<archive>
-						<manifestEntries>
-							<Plugin-Class>${plugin.class}</Plugin-Class>
-							<Plugin-Dependencies>${plugin.dependencies}</Plugin-Dependencies>
-							<Plugin-Id>${plugin.id}</Plugin-Id>
-							<Plugin-Provider>${plugin.provider}</Plugin-Provider>
-							<Plugin-Version>${plugin.version}</Plugin-Version>
-							<Plugin-Requires>${plugin.requires.runtime}</Plugin-Requires>
-						</manifestEntries>
-					</archive>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-assembly-plugin</artifactId>
-				<version>2.6</version>
-				<configuration>
-					<descriptorRefs>
-						<descriptorRef>jar-with-dependencies</descriptorRef>
-					</descriptorRefs>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
