--- irida-snippy-pipeline-plugin/src/main/resources/workflows/0.1.1/messages_en.properties
@@ -1,20 +1,0 @@
-#Pipeline Info Properties
-#Thu Mar 21 10:42:46 SAST 2019
-workflow.SNIPPY_BACTERIAL_VARIANT_CALLER.title=SnippyBVCPipeline Pipeline
-pipeline.parameters.modal-title.snippybvcpipeline=SnippyBVCPipeline Pipeline Parameters
-pipeline.title.SnippyBVCPipeline=Pipelines - SnippyBVCPipeline
-workflow.SNIPPY_BACTERIAL_VARIANT_CALLER.description=Snippy finds SNPs between a haploid reference genome and your NGS sequence reads. It will find both substitutions (snps) and insertions/deletions (indels). It will use as many CPUs as you can give it on a single computer (tested to 64 cores).
-pipeline.h1.SnippyBVCPipeline=SnippyBVCPipeline Pipeline
-#Tool Parameters - Tool: confil - Workflow Step #: 2
-#Thu Mar 21 10:42:46 SAST 2019
-pipeline.parameters.snippybvcpipeline.confil-2-input_type_conditional.input_type=confil-2-input_type_conditional.input_type
-pipeline.parameters.snippybvcpipeline.confil-2-cutoff=confil-2-cutoff
-#Tool Parameters - Tool: snippy - Workflow Step #: 4
-#Thu Mar 21 10:42:46 SAST 2019
-pipeline.parameters.snippybvcpipeline.snippy-4-adv.mapqual=snippy-4-adv.mapqual
-pipeline.parameters.snippybvcpipeline.snippy-4-adv.rgid=snippy-4-adv.rgid
-pipeline.parameters.snippybvcpipeline.snippy-4-fastq_input.fastq_input_selector=snippy-4-fastq_input.fastq_input_selector
-pipeline.parameters.snippybvcpipeline.snippy-4-adv.mincov=snippy-4-adv.mincov
-pipeline.parameters.snippybvcpipeline.snippy-4-adv.minfrac=snippy-4-adv.minfrac
-pipeline.parameters.snippybvcpipeline.snippy-4-adv.bwaopt=snippy-4-adv.bwaopt
-workflow.label.share-analysis-samples.READ_INFO=Save sequence read information to Project Line List Metadata
