+++ sanbi-bvc-pipeline-plugin/src/main/resources/workflows/0.1.0/irida_workflow_structure.ga
