+++ snippy-pipeline-plugin/src/main/resources/workflows/0.1.1/irida_workflow.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<iridaWorkflow>
+  <id>26385d7e-7665-4140-9253-e8d721216f59</id>
+  <name>Snippy</name>
+  <version>0.1.1</version>
+  <analysisType>VARIANT_CALLING</analysisType>
+  <inputs>
+    <sequenceReadsPaired>Input Collection</sequenceReadsPaired>
+    <reference>Reference File</reference>
+    <requiresSingleSample>true</requiresSingleSample>
+  </inputs>
+  <parameters>
+    <parameter name="snippy-2-adv.bwaopt" defaultValue="">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="adv.bwaopt" />
+    </parameter>
+    <parameter name="snippy-2-adv.mapqual" defaultValue="60">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="adv.mapqual" />
+    </parameter>
+    <parameter name="snippy-2-adv.mincov" defaultValue="10">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="adv.mincov" />
+    </parameter>
+    <parameter name="snippy-2-adv.minfrac" defaultValue="0.9">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="adv.minfrac" />
+    </parameter>
+    <parameter name="snippy-2-adv.minqual" defaultValue="100.0">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="adv.minqual" />
+    </parameter>
+    <parameter name="snippy-2-adv.rename_cons" defaultValue="false">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="adv.rename_cons" />
+    </parameter>
+    <parameter name="snippy-2-adv.rgid" defaultValue="">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="adv.rgid" />
+    </parameter>
+    <parameter name="snippy-2-fastq_input.fastq_input_selector" defaultValue="paired_collection">
+      <toolParameter toolId="toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1" parameterName="fastq_input.fastq_input_selector" />
+    </parameter>
+  </parameters>
+  <outputs>
+    <output name="snippy_core" fileName="snippy_core" />
+    <output name="consensus_fasta" fileName="consensus_fasta" />
+    <output name="aligned_fasta" fileName="aligned_fasta" />
+    <output name="snps_gff_file" fileName="snps_gff_file" />
+    <output name="snps_summary" fileName="snps_summary" />
+    <output name="snps_vcf_file" fileName="snps_vcf_file" />
+    <output name="log_file" fileName="log_file" />
+    <output name="mapped_reads" fileName="mapped_reads" />
+    <output name="mapping_depth" fileName="mapping_depth" />
+    <output name="snps_table" fileName="snps_table" />
+  </outputs>
+  <toolRepositories>
+    <repository>
+      <name>snippy</name>
+      <owner>iuc</owner>
+      <url>https://toolshed.g2.bx.psu.edu</url>
+      <revision>776ebd1239da</revision>
+    </repository>
+  </toolRepositories>
+</iridaWorkflow>
