+++ snippy-pipeline-plugin/src/main/resources/workflows/0.1.1/irida_workflow_structure.ga
@@ -1,0 +1,1 @@
+{"uuid": "5c8161f6-d3ca-4e77-9bc2-57fab9bd3ac8", "tags": [], "format-version": "0.1", "name": "Copy of Snippy Variant Calling v2", "version": 1, "steps": {"0": {"tool_id": null, "tool_version": null, "outputs": [], "workflow_outputs": [{"output_name": "output", "uuid": "29aedf1b-28bf-4889-86db-d9efbd38bda3", "label": null}], "input_connections": {}, "tool_state": "{\"name\": \"reference\"}", "id": 0, "uuid": "8d183e0a-bb11-46c0-9b34-576e194e5ae7", "errors": null, "name": "Input dataset", "label": "Reference File", "inputs": [{"name": "reference", "description": ""}], "position": {"top": 230, "left": 261}, "annotation": "", "content_id": null, "type": "data_input"}, "1": {"tool_id": null, "tool_version": null, "outputs": [], "workflow_outputs": [{"output_name": "output", "uuid": "757baf52-61e3-4a8b-b8f1-118b4a5155e7", "label": null}], "input_connections": {}, "tool_state": "{\"collection_type\": \"list:paired\"}", "id": 1, "uuid": "f05ae535-aefc-4e1f-8d3d-da85e0470d71", "errors": null, "name": "Input dataset collection", "label": "Input Collection", "inputs": [], "position": {"top": 410, "left": 200}, "annotation": "", "content_id": null, "type": "data_collection_input"}, "2": {"tool_id": "toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1", "tool_version": "4.3.6+galaxy1", "outputs": [{"type": "vcf", "name": "snpvcf"}, {"type": "gff3", "name": "snpgff"}, {"type": "tabular", "name": "snptab"}, {"type": "tabular", "name": "snpsum"}, {"type": "txt", "name": "snplog"}, {"type": "fasta", "name": "snpalign"}, {"type": "fasta", "name": "snpconsensus"}, {"type": "tabular", "name": "snpsdepth"}, {"type": "bam", "name": "snpsbam"}, {"type": "zip", "name": "outdir"}], "workflow_outputs": [{"output_name": "snplog", "uuid": "c5876b27-6579-4aa3-8d7a-4dcee116cabb", "label": "log_file"}, {"output_name": "snptab", "uuid": "30ac1ad8-95c5-4618-b01f-c026a6c3f900", "label": "snps_table"}, {"output_name": "snpvcf", "uuid": "24498a00-2532-4160-90d7-d0e3214f1301", "label": "snps_vcf_file "}, {"output_name": "snpgff", "uuid": "cba676c6-0ce7-4726-8e01-684c355f07a8", "label": "snps_gff_file"}, {"output_name": "snpconsensus", "uuid": "a84f7790-16c9-4f58-8832-a34430ceb885", "label": "consensus_fasta"}, {"output_name": "snpsbam", "uuid": "4df93d52-03d7-4ead-96c6-e66ed341e262", "label": "mapped_reads"}, {"output_name": "snpalign", "uuid": "c7bdfb66-9629-40cb-a93b-98d80cce5431", "label": "aligned_fasta"}, {"output_name": "snpsum", "uuid": "4c8d144d-7ad5-4651-a4cc-b1f3dc6797ed", "label": "snps_summary"}, {"output_name": "snpsdepth", "uuid": "a022bd3c-2a3a-42f3-bcb8-731b3df3a2d8", "label": "mapping_depth"}, {"output_name": "outdir", "uuid": "4e62ab0d-d35e-4535-b800-d22560f73d74", "label": "snippy_core"}], "input_connections": {"ref": {"output_name": "output", "id": 0}, "fastq_input|fastq_input": {"output_name": "output", "id": 1}}, "tool_state": "{\"adv\": \"{\\\"bwaopt\\\": \\\"\\\", \\\"mapqual\\\": \\\"60\\\", \\\"mincov\\\": \\\"10\\\", \\\"minfrac\\\": \\\"0.9\\\", \\\"minqual\\\": \\\"100.0\\\", \\\"rename_cons\\\": \\\"false\\\", \\\"rgid\\\": \\\"\\\"}\", \"__page__\": null, \"outputs\": \"[\\\"outvcf\\\", \\\"outgff\\\", \\\"outtab\\\", \\\"outsum\\\", \\\"outlog\\\", \\\"outaln\\\", \\\"outcon\\\", \\\"outdep\\\", \\\"outbam\\\", \\\"outzip\\\"]\", \"__rerun_remap_job_id__\": null, \"fastq_input\": \"{\\\"__current_case__\\\": 2, \\\"fastq_input\\\": {\\\"__class__\\\": \\\"ConnectedValue\\\"}, \\\"fastq_input_selector\\\": \\\"paired_collection\\\"}\", \"ref\": \"{\\\"__class__\\\": \\\"ConnectedValue\\\"}\"}", "id": 2, "tool_shed_repository": {"owner": "iuc", "changeset_revision": "776ebd1239da", "name": "snippy", "tool_shed": "toolshed.g2.bx.psu.edu"}, "uuid": "b48c44a3-8bd4-43ae-af2f-88df5bea01ff", "errors": null, "name": "snippy", "post_job_actions": {"RenameDatasetActionsnpgff": {"output_name": "snpgff", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "snps_gff_file"}}, "RenameDatasetActionsnpsum": {"output_name": "snpsum", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "snps_summary"}}, "RenameDatasetActionsnpvcf": {"output_name": "snpvcf", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "snps_vcf_file"}}, "RenameDatasetActionsnptab": {"output_name": "snptab", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "snps_table"}}, "RenameDatasetActionsnplog": {"output_name": "snplog", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "log_file"}}, "RenameDatasetActionsnpconsensus": {"output_name": "snpconsensus", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "consensus_fasta"}}, "RenameDatasetActionoutdir": {"output_name": "outdir", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "snippy_core"}}, "RenameDatasetActionsnpalign": {"output_name": "snpalign", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "aligned_fasta"}}, "RenameDatasetActionsnpsbam": {"output_name": "snpsbam", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "mapped_reads"}}, "RenameDatasetActionsnpsdepth": {"output_name": "snpsdepth", "action_type": "RenameDatasetAction", "action_arguments": {"newname": "mapping_depth"}}}, "label": "snippy", "inputs": [], "position": {"top": 193, "left": 701}, "annotation": "", "content_id": "toolshed.g2.bx.psu.edu/repos/iuc/snippy/snippy/4.3.6+galaxy1", "type": "tool"}}, "annotation": "", "a_galaxy_workflow": "true"}
