+++ snippy-pipeline-plugin/src/main/resources/workflows/0.1.1/messages_en.properties
@@ -1,0 +1,19 @@
+#Pipeline Info Properties
+#Wed Mar 27 15:15:23 SAST 2019
+pipeline.h1.SnippyPipeline=Snippy Pipeline
+workflow.VARIANT_CALLING.title=Snippy Pipeline
+pipeline.parameters.modal-title.snippypipeline=Snippy Pipeline Parameters
+workflow.VARIANT_CALLING.description=Snippy finds SNPs between a haploid reference genome and your NGS sequence reads. It will find both substitutions (snps) and insertions/deletions (indels).
+pipeline.title.SnippyPipeline=Pipelines - Snippy
+#Tool Parameters - Tool: snippy - Workflow Step #: 2
+#Wed Mar 27 15:15:23 SAST 2019
+pipeline.parameters.snippypipeline.snippy-2-adv.rgid=snippy-2-adv.rgid
+pipeline.parameters.snippypipeline.snippy-2-adv.minfrac=snippy-2-adv.minfrac
+pipeline.parameters.snippypipeline.snippy-2-adv.rename_cons=snippy-2-adv.rename_cons
+pipeline.parameters.snippypipeline.snippy-2-adv.mincov=snippy-2-adv.mincov
+pipeline.parameters.snippypipeline.snippy-2-adv.mapqual=snippy-2-adv.mapqual
+pipeline.parameters.snippypipeline.snippy-2-adv.minqual=snippy-2-adv.minqual
+pipeline.parameters.snippypipeline.snippy-2-fastq_input.fastq_input_selector=snippy-2-fastq_input.fastq_input_selector
+pipeline.parameters.snippypipeline.snippy-2-adv.bwaopt=snippy-2-adv.bwaopt
+workflow.label.share-analysis-samples.VARIANT_CALLING=Save sequence read information to Project Line List Metadata
+
